refactor(stage-pages): share polaroid page
