+++ packages/stage-pages/README.md
@@ -1,0 +1,23 @@
+# Stage Pages
+
+`@proj-airi/stage-pages` contains route-level Vue pages that multiple AIRI stage applications share.
+
+## Use
+
+Add a page under `src/pages`. Each stage application scans this directory through its Vite router configuration.
+
+Import a shared page through the package boundary when code needs the component directly:
+
+```ts
+import PolaroidPage from '@proj-airi/stage-pages/pages/devtools/polaroid.vue'
+```
+
+## When to use this package
+
+Use this package when two or more stage applications need the same page behavior and route structure.
+
+## When not to use this package
+
+Keep application-specific pages in the owning application. Put reusable business components in `@proj-airi/stage-ui`.
+
+Put primitive UI components in `@proj-airi/ui`. Put shared layouts in `@proj-airi/stage-layouts`.
