+++ packages/stage-pages/src/pages/devtools/polaroid.vue
